requirements.txt
asgiref==3.5.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
coverage==7.2.7
Django==4.1.2
django-cors-headers==3.13.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
et-xmlfile==1.1.0
exceptiongroup==1.1.1
idna==3.4
importlib-resources==5.12.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.17.3
openpyxl==3.0.10
packaging==23.1
pkgutil_resolve_name==1.3.10
pluggy==1.2.0
PyJWT==2.6.0
pyrsistent==0.19.3
pytest==7.3.1
pytest-cov==4.1.0
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2022.4
PyYAML==6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
uritemplate==4.1.1
urllib3==2.0.3
zipp==3.15.0
psycopg2-binary==2.9.6
django-environ==0.10.0
gunicorn==20.1.0
flake8==6.1.0
pre-commit==3.5.0
factory-boy==3.2.1
Faker==19.1.0
amqp==5.2.0
asgiref==3.5.2
attrs==23.1.0
backports.zoneinfo==0.2.1
billiard==4.2.0
celery==5.3.6
certifi==2023.5.7
cfgv==3.4.0
charset-normalizer==3.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.2.7
distlib==0.3.7
Django==4.1.2
django-cors-headers==3.13.0
django-environ==0.10.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
et-xmlfile==1.1.0
exceptiongroup==1.1.1
filelock==3.13.1
flake8==6.1.0
gunicorn==20.1.0
identify==2.5.31
idna==3.4
importlib-resources==5.12.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.17.3
kombu==5.3.5
mccabe==0.7.0
nodeenv==1.8.0
openpyxl==3.0.10
packaging==23.1
pkgutil_resolve_name==1.3.10
platformdirs==3.11.0
pluggy==1.2.0
pre-commit==3.5.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.6
pycodestyle==2.11.1
pyflakes==3.1.0
PyJWT==2.6.0
pyrsistent==0.19.3
pytest==7.3.1
pytest-cov==4.1.0
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2022.4
PyYAML==6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.10.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==2.0.3
vine==5.1.0
virtualenv==20.24.6
wcwidth==0.2.13
zipp==3.15.0
amqp==5.2.0
asgiref==3.5.2
async-timeout==4.0.3
attrs==23.1.0
backports.zoneinfo==0.2.1
billiard==4.2.0
celery==5.3.6
certifi==2023.5.7
cfgv==3.4.0
charset-normalizer==3.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.2.7
distlib==0.3.7
Django==4.1.2
django-cors-headers==3.13.0
django-environ==0.10.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
et-xmlfile==1.1.0
exceptiongroup==1.1.1
filelock==3.13.1
flake8==6.1.0
gunicorn==20.1.0
identify==2.5.31
idna==3.4
importlib-resources==5.12.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.17.3
kombu==5.3.5
mccabe==0.7.0
nodeenv==1.8.0
openpyxl==3.0.10
packaging==23.1
pkgutil_resolve_name==1.3.10
platformdirs==3.11.0
pluggy==1.2.0
pre-commit==3.5.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.6
pycodestyle==2.11.1
pyflakes==3.1.0
PyJWT==2.6.0
pyrsistent==0.19.3
pytest==7.3.1
pytest-cov==4.1.0
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2022.4
PyYAML==6.0
redis==5.0.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.10.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==2.0.3
vine==5.1.0
virtualenv==20.24.6
wcwidth==0.2.13
zipp==3.15.0
amqp==5.2.0
asgiref==3.5.2
async-timeout==4.0.3
attrs==23.1.0
backports.zoneinfo==0.2.1
billiard==4.2.0
black==24.2.0
celery==5.3.6
certifi==2023.5.7
cfgv==3.4.0
charset-normalizer==3.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.2.7
distlib==0.3.7
Django==4.1.2
django-cors-headers==3.13.0
django-environ==0.10.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
et-xmlfile==1.1.0
exceptiongroup==1.1.1
Faker==24.3.0
filelock==3.13.1
flake8==6.1.0
gunicorn==20.1.0
identify==2.5.31
idna==3.4
importlib-resources==5.12.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.17.3
kombu==5.3.5
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.8.0
openpyxl==3.0.10
packaging==23.1
pathspec==0.12.1
pkgutil_resolve_name==1.3.10
platformdirs==3.11.0
pluggy==1.2.0
pre-commit==3.5.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.6
pycodestyle==2.11.1
pyflakes==3.1.0
PyJWT==2.6.0
pyrsistent==0.19.3
pytest==7.3.1
pytest-cov==4.1.0
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2022.4
PyYAML==6.0
redis==5.0.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.10.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==2.0.3
vine==5.1.0
virtualenv==20.24.6
wcwidth==0.2.13
zipp==3.15.0
